#3 graph features * move all packages into the backend package
