+++ backend/__init__.py
@@ -1,1 +1,2 @@
 from .controller import Controller
+from .graph import collaborative, hero_comic
